feat(tags): Add tag category label getter and centralize consts
